@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 version = "0.0.1"
 name = "easywhisper"
 
-requires-python = ">= 3.9"
+requires-python = ">= 3.10"
 
 dependencies = [
   "torch>=2.1.2",
@@ -21,7 +21,7 @@ readme = "README.md"
 line-length = 99
 indent-width = 4
 
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 # Allow unused variables when underscore-prefixed.
