@@ -12,7 +12,16 @@ def transcribe(
     model: WhisperForConditionalGeneration,
     processor: WhisperProcessor,
     file_dataloader: torch.utils.data.DataLoader,
+    language: str | None = None,
+    batch_size: int = 4,
+    num_workers: int = 2,
+    prefetch_factor: int = 2,
+    max_length: int = 250,
+    num_beams: int = 5,
+    length_penalty: float = 1.0,
+    repetition_penalty: float = 1.0,
     output_dir: str = "output/transcriptions",
+    device: str = "cuda",
 ):
     for features in file_dataloader:
         slice_dataset = features[0]["dataset"]
@@ -21,22 +30,26 @@ def transcribe(
 
         feature_dataloader = torch.utils.data.DataLoader(
             slice_dataset,
-            batch_size=4,
-            num_workers=2,
-            prefetch_factor=2,
+            batch_size=batch_size,
+            num_workers=num_workers,
+            prefetch_factor=prefetch_factor,
             collate_fn=transcribe_collate_fn,
         )
 
         for batch in feature_dataloader:
             with torch.inference_mode():
-                batch = batch["features"].to("cuda").half()
+                batch = batch["features"].to(device).half()
                 predicted_ids = model.generate(
                     batch,
                     return_dict_in_generate=True,
                     task="transcribe",
-                    language="sv",
-                    output_scores=True,
-                    max_length=250,
+                    language=language,
+                    output_scores=False,
+                    max_length=max_length,
+                    num_beams=num_beams,
+                    repetition_penalty=repetition_penalty,
+                    length_penalty=length_penalty,
+                    early_stopping=True,
                 )
 
                 transcription = processor.batch_decode(
